backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.6.0
supabase==2.0.2
kafka-python>=2.0.0
python-dotenv>=1.0.0
pypdf>=3.17.0
pyyaml>=6.0.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
httpx>=0.24.0,<0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
websockets==10.4
nltk>=3.8.0 